HIGH - Uncoded feature packs preventing the error import from proceeding to the next step
MEDIUM - Import not throwing an error if we have missing / extra columns or they are in an incorrect order. I found this when I tried to import an old spreadsheet as opposed to the new v4 with the derivative description column.
LOW - Erroneous empty red error message on import complete.